// apps/server/src/app/graphql/enums/column-view-type.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum ColumnViewTypeEnum {
  MATCHES_CREATED_AT = 'MATCHES_CREATED_AT',
  MATCHES_UPDATED_AT = 'MATCHES_UPDATED_AT',
  MATCHES_INGREDIENT_TEXT = 'MATCHES_INGREDIENT_TEXT',
  MATCHES_STATUS = 'MATCHES_STATUS',
  MATCHES_SELECTED_FOOD_MATCH_ID = 'MATCHES_SELECTED_FOOD_MATCH_ID',

  RECIPES_DESCRIPTION = 'RECIPES_DESCRIPTION',
  RECIPES_SERVINGS_MAX = 'RECIPES_SERVINGS_MAX',
  RECIPES_SERVINGS_MIN = 'RECIPES_SERVINGS_MIN',
  RECIPES_SERVINGS_TEXT = 'RECIPES_SERVINGS_TEXT',
  RECIPES_SOURCE_URL = 'RECIPES_SOURCE_URL',
  RECIPES_TITLE = 'RECIPES_TITLE',
  RECIPES_CREATED_AT = 'RECIPES_CREATED_AT',
  RECIPES_UPDATED_AT = 'RECIPES_UPDATED_AT',
}

registerEnumType(ColumnViewTypeEnum, {
  name: 'ColumnViewType',
});